+++ test/deno_entry.js
@@ -2,6 +2,7 @@ import "./deno_shim.js";
 
 // Import all tests to run them in one Deno process with the shim
 import "./GitBlob.test.js";
+import "./GitCommitFlow.test.js";
 import "./GitRef.test.js";
 import "./GitSha.test.js";
 import "./ShellRunner.test.js";
@@ -18,5 +19,6 @@ import "./domain/services/CommandSanitizer.test.js";
 import "./domain/services/EnvironmentPolicy.test.js";
 import "./domain/services/GitCommandBuilder.test.js";
 import "./domain/services/GitErrorClassifier.test.js";
+import "./domain/services/GitPersistenceService.test.js";
 import "./domain/value-objects/GitFileMode.test.js";
 import "./domain/value-objects/GitObjectType.test.js";